+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fix `alwaysInstallModules = true` not triggers pipInstall for non-strict requirements file (#94)
   (required for case when requirements file links other files, which changes are not tracked) 
+* Add requirements file references (-r file) support under strict mode (#94) 
 
 ### 3.0.0 (2022-10-22)
 * (breaking) Drop gradle 5.0-5.2 support (minimum required gradle is 5.3)